RENAME #17 IInventory
