+++ src/logic/IInventory.ts
