+++ src/logic/logic.test.ts
@@ -6,8 +6,8 @@ import * as sinon from "sinon";
 import * as sinonChai from "sinon-chai";
 import { CityName } from "./CityName";
 import { doNothing } from "./doNothing";
-import { IInventory } from "./i-inventory";
 import { ICity } from "./ICity";
+import { IInventory } from "./IInventory";
 import { Logic } from "./logic";
 import { WareType } from "./wareType";
